+++ app/src/main/java/com/internetshop/model/Goods.java
@@ -27,7 +27,7 @@ public class Goods {
     private String description;
     private String img;
     private int salesCounter;
-    private int rating;
+    private double rating;
     private Category category;
     private Rule rule;
     private String date;
@@ -50,7 +50,7 @@ public class Goods {
         this.rule = rule;
     }
 
-    public Goods(int id, String name, float price, int numberOfPlayers, float duration, int amount, int visible, String description, String img, int salesCounter, int rating, Category category, Rule rule, String date) {
+    public Goods(int id, String name, float price, int numberOfPlayers, float duration, int amount, int visible, String description, String img, int salesCounter, double rating, Category category, Rule rule, String date) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -163,11 +163,11 @@ public class Goods {
         this.salesCounter = salesCounter;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
